refactor tapnet
