+++ aeon/classification/deep_learning/__init__.py
@@ -25,5 +25,5 @@ from aeon.classification.deep_learning._lite_time import (
 )
 from aeon.classification.deep_learning._mlp import MLPClassifier
 from aeon.classification.deep_learning._resnet import ResNetClassifier
+from aeon.classification.deep_learning._tapnet import TapNetClassifier
 from aeon.classification.deep_learning.base import BaseDeepClassifier
-from aeon.classification.deep_learning.tapnet import TapNetClassifier
