+++ aeon/classification/deep_learning/_tapnet.py
